feat: add TabCountButton component to MobileBrowserBottomBar and simplify bottom bar logic
- Introduced a new TabCountButton component to display the number of tabs and handle tab list navigation.
- Removed redundant tab count logic from MobileBrowserBottomBar and updated the bottom bar's background color.
- Simplified the displayBottomBar logic in MobileBrowser for improved clarity.
